+++ src/main/java/learn/springang/umsportal/exceptions/domain/EmailExistsException.java
@@ -1,0 +1,7 @@
+package learn.springang.umsportal.exceptions.domain;
+
+public class EmailExistsException extends Exception {
+    public EmailExistsException(String message) {
+        super(message);
+    }
+}