+++ src/main/java/learn/springang/umsportal/exceptions/domain/EmailNotFoundException.java
@@ -1,0 +1,7 @@
+package learn.springang.umsportal.exceptions.domain;
+
+public class EmailNotFoundException extends Exception {
+    public EmailNotFoundException(String message) {
+        super(message);
+    }
+}