+++ src/main/java/learn/springang/umsportal/exceptions/domain/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package learn.springang.umsportal.exceptions.domain;
+
+public class UserNotFoundException extends Exception {
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}