+++ src/main/java/learn/springang/umsportal/exceptions/domain/UsernameExistsException.java
@@ -1,0 +1,7 @@
+package learn.springang.umsportal.exceptions.domain;
+
+public class UsernameExistsException extends Exception {
+    public UsernameExistsException(String message) {
+        super(message);
+    }
+}